split up bounce env
